Make value_size a size_t as I should have done in the first place probably

+++ src/randomPlantedForest.cpp
@@ -511,11 +511,11 @@ template <typename VT>
 std::vector<std::vector<VT>> transpose(std::vector<std::vector<VT>>& mat){
 
   if(mat.size()<=0) throw std::invalid_argument("Matrix is empty.");
-  int value_size = mat[0].size();
+  size_t value_size = mat[0].size();
 
   std::vector<std::vector<VT>> columns(value_size);
   for(auto vec: mat){
-    for(int n_col=0; n_col<value_size; ++n_col){
+    for(size_t n_col=0; n_col<value_size; ++n_col){
       columns[n_col].push_back(vec[n_col]);
     }
   }
@@ -717,7 +717,7 @@ int n_splits;                               /**< Number of performed splits for 
 std::vector<int> n_leaves;                  /**< */
 double t_try = 0.4;                         /**< */
 int split_try = 10;                         /**< */
-int value_size = 1;
+size_t value_size = 1;
 int feature_size = 0;                       /**< Number of feature dimension in X */
 int sample_size = 0;                        /**< Number of samples of X */
 bool purify_forest = 0;                     /**< Whether the forest should be purified */
@@ -748,7 +748,7 @@ void RandomPlantedForest::L2_loss(rpf::Split& split){
   split.M_b = split.sum_b / split.I_b.size();
 
   split.min_sum = 0;
-  for(int p=0; p<value_size; ++p){
+  for(size_t p=0; p<value_size; ++p){
     split.min_sum  += - 2 * split.M_s[p] * split.sum_s[p] + split.I_s.size() * pow(split.M_s[p], 2);
     split.min_sum  += - 2 * split.M_b[p] * split.sum_b[p] + split.I_b.size() * pow(split.M_b[p], 2);
   }
@@ -1534,11 +1534,11 @@ void RandomPlantedForest::purify_1(){
                 Leaf new_leaf = curr_leaf; // initialize intervals with first leaf
                 new_leaf.intervals[feature_dim-1].first = lower_bounds[feature_dim-1];
                 new_leaf.intervals[feature_dim-1].second = upper_bounds[feature_dim-1];
-                for(int i=0; i<value_size; ++i) new_leaf.value[i] = -curr_leaf.value[i] * multiplier; // update value of new leaf
+                for(size_t i=0; i<value_size; ++i) new_leaf.value[i] = -curr_leaf.value[i] * multiplier; // update value of new leaf
 
                 // append new leaf
                 if(!leafExists(new_leaf.intervals, curr_tree)) curr_tree->leaves.push_back(new_leaf);
-                for(int i=0; i<value_size; ++i) new_leaf.value[i] = curr_leaf.value[i] * multiplier; // update value of new leaf
+                for(size_t i=0; i<value_size; ++i) new_leaf.value[i] = curr_leaf.value[i] * multiplier; // update value of new leaf
                 if(!leafExists(new_leaf.intervals, tree)) tree->leaves.push_back(new_leaf);
               }
             }
@@ -2538,7 +2538,7 @@ void ClassificationRPF::L1_loss(rpf::Split& split){
   split.M_s = split.sum_s / split.I_s.size();
   split.M_b = split.sum_b / split.I_b.size();
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       split.min_sum += std::fabs((*split.Y)[individual][p] - split.M_s[p]) - std::fabs((*split.Y)[individual][p]);
     }
@@ -2553,7 +2553,7 @@ void ClassificationRPF::median_loss(rpf::Split& split){
   split.M_s = calcMedian(*split.Y, split.I_s);
   split.M_b = calcMedian(*split.Y, split.I_b);
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       split.min_sum += std::fabs((*split.Y)[individual][p] - split.M_s[p]) - std::fabs((*split.Y)[individual][p]);
     }
@@ -2585,7 +2585,7 @@ void ClassificationRPF::logit_loss(rpf::Split& split){
 
   std::vector<std::vector<double>> W = *split.W, W_new = *split.W;
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       W[individual][p] = exp(W[individual][p]);
       W_new[individual][p] = exp(W_new[individual][p] + log(M_s[p] / M_sp) - W_s_mean[p]);
@@ -2596,7 +2596,7 @@ void ClassificationRPF::logit_loss(rpf::Split& split){
     }
   }
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       split.min_sum += (*split.Y)[individual][p] * log(W[individual][p] / (1 + std::accumulate(W[individual].begin(), W[individual].end(), 0.0)) ); // ~ R_old
       split.min_sum -= (*split.Y)[individual][p] * log(W_new[individual][p] / (1 + std::accumulate(W_new[individual].begin(), W_new[individual].end(), 0.0)) ); // ~ R_new
@@ -2644,7 +2644,7 @@ void ClassificationRPF::logit_loss_2(rpf::Split& split){
 
   std::vector<std::vector<double>> W = *split.W, W_new = *split.W;
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       W[individual][p] = exp(W[individual][p]);
       W_new[individual][p] = exp(W_new[individual][p] + log(M_s[p] / M_s2[p]) - W_s_mean[p]);
@@ -2655,7 +2655,7 @@ void ClassificationRPF::logit_loss_2(rpf::Split& split){
     }
   }
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       split.min_sum += (*split.Y)[individual][p] * log(W[individual][p] / (1 + W[individual][p] )); // ~ R_old
       split.min_sum -= (*split.Y)[individual][p] * log(W_new[individual][p] / (1 + W_new[individual][p] )); // ~ R_new
@@ -2705,7 +2705,7 @@ void ClassificationRPF::logit_loss_3(rpf::Split& split){
   //std::vector<std::vector<double>> Y_s = split.Y_s;
   //std::vector<std::vector<double>> Y_b = split.Y_b;
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       W_new[individual][p] = W_new[individual][p] + M_s[p] - sum_s - W_s_mean[p];
     }
@@ -2755,7 +2755,7 @@ void ClassificationRPF::logit_loss_3(rpf::Split& split){
    Y_b = transpose(Y_b);
    */
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       W[individual][p] = exp(W[individual][p]);
       W_new[individual][p] = exp(W_new[individual][p]);
@@ -2766,7 +2766,7 @@ void ClassificationRPF::logit_loss_3(rpf::Split& split){
     }
   }
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       split.min_sum += (*split.Y)[individual][p] * log(W[individual][p] / (std::accumulate(W[individual].begin(), W[individual].end(), 0.0)) ); // ~ R_old
       split.min_sum -= (*split.Y)[individual][p] * log(W_new[individual][p] / (std::accumulate(W_new[individual].begin(), W_new[individual].end(), 0.0)) ); // ~ R_new
@@ -2805,7 +2805,7 @@ void ClassificationRPF::logit_loss_4(rpf::Split& split){
 
   std::vector<std::vector<double>> W = *split.W, W_new = *split.W;
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       W[individual][p] = exp(W[individual][p]);
       W_new[individual][p] = exp(W_new[individual][p] + log(M_s[p] / M_s2[p]) - W_s_mean[p]);
@@ -2816,7 +2816,7 @@ void ClassificationRPF::logit_loss_4(rpf::Split& split){
     }
   }
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       split.min_sum += (*split.Y)[individual][p] * log(W[individual][p] / (1 + W[individual][p] )); // ~ R_old
       split.min_sum -= (*split.Y)[individual][p] * log(W_new[individual][p] / (1 + W_new[individual][p] )); // ~ R_new
@@ -2842,7 +2842,7 @@ void ClassificationRPF::exponential_loss(rpf::Split& split){
   std::vector<double> sum_s(value_size, 0);
   std::vector<double> sum_b(value_size, 0);
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       W_s_sum[p] += (*split.W)[individual][p];
     }
@@ -2869,7 +2869,7 @@ void ClassificationRPF::exponential_loss(rpf::Split& split){
   double sum_sp = std::min(std::max(delta, split.M_sp), 1-delta);
   double sum_bp = std::min(std::max(delta, split.M_bp), 1-delta);
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
     for(auto individual: split.I_s){
       split.min_sum += (*split.W)[individual][p] * exp(-0.5 * (*split.Y)[individual][p] * log(sum_s[p] / sum_sp));
     }
@@ -2896,7 +2896,7 @@ void ClassificationRPF::exponential_loss_2(rpf::Split& split){
   std::vector<double> sum_s2(value_size, 0);
   std::vector<double> sum_b2(value_size, 0);
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
 
     for(auto individual: split.I_s){
       W_s_sum[p] += (*split.W)[individual][p];
@@ -2922,7 +2922,7 @@ void ClassificationRPF::exponential_loss_2(rpf::Split& split){
     sum_b[p] = std::max(delta, sum_b[p]);
   }
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
 
     for(auto individual: split.I_s){
       split.min_sum += (*split.W)[individual][p] * exp(-0.5 * (*split.Y)[individual][p] * log(sum_s[p] / sum_s2[p]));
@@ -2950,7 +2950,7 @@ void ClassificationRPF::exponential_loss_3(rpf::Split& split){
   std::vector<double> sum_s(value_size, 0);
   std::vector<double> sum_b(value_size, 0);
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
 
     for(auto individual: split.I_s){
       W_s_sum[p] += (*split.W)[individual][p];
@@ -2989,7 +2989,7 @@ void ClassificationRPF::exponential_loss_3(rpf::Split& split){
   sum_sp = sum_sp / (sum_s.size() + 1);
   sum_bp = sum_bp / (sum_b.size() + 1);
 
-  for(int p=0; p < static_cast<int>(value_size); ++p){
+  for(size_t p=0; p < value_size; ++p){
 
     for(auto individual: split.I_s){
       split.min_sum += (*split.W)[individual][p] * exp(-0.5 * (*split.Y)[individual][p] * (sum_s[p] - sum_sp));
@@ -3346,7 +3346,7 @@ void ClassificationRPF::create_tree_family(std::vector<Leaf> initial_leaves, siz
         std::vector<double> W_s_mean = calcMean(*curr_split.W, curr_split.I_s);
         std::vector<double> W_b_mean = calcMean(*curr_split.W, curr_split.I_b);
 
-        for(unsigned int p=0; p < static_cast<unsigned int>(value_size); ++p){
+        for(size_t p=0; p < value_size; ++p){
           update_s[p] = log(M_s[p] / M_sp) - W_s_mean[p];
           update_b[p] = log(M_b[p] / M_bp) - W_b_mean[p];
         }
@@ -3378,7 +3378,7 @@ void ClassificationRPF::create_tree_family(std::vector<Leaf> initial_leaves, siz
         std::vector<double> W_s_mean = calcMean(*curr_split.W, curr_split.I_s);
         std::vector<double> W_b_mean = calcMean(*curr_split.W, curr_split.I_b);
 
-        for(int p=0; p < static_cast<int>(value_size); ++p){
+        for(size_t p=0; p < value_size; ++p){
           update_s[p] = log(M_s[p] / M_s2[p]) - W_s_mean[p];
           update_b[p] = log(M_b[p] / M_b2[p]) - W_b_mean[p];
         }
@@ -3448,7 +3448,7 @@ void ClassificationRPF::create_tree_family(std::vector<Leaf> initial_leaves, siz
         std::vector<double> W_s_mean = calcMean(*curr_split.W, curr_split.I_s);
         std::vector<double> W_b_mean = calcMean(*curr_split.W, curr_split.I_b);
 
-        for(int p=0; p < static_cast<int>(value_size); ++p){
+        for(size_t p=0; p < value_size; ++p){
           update_s[p] = log(M_s[p] / M_s2[p]) - W_s_mean[p];
           update_b[p] = log(M_b[p] / M_b2[p]) - W_b_mean[p];
         }
@@ -3504,13 +3504,13 @@ void ClassificationRPF::create_tree_family(std::vector<Leaf> initial_leaves, siz
         std::for_each(sum_s2.begin(), sum_s2.end(), [this](double &S) { S = std::max(epsilon, 1 - S); });
         std::for_each(sum_b2.begin(), sum_b2.end(), [this](double &S) { S = std::max(epsilon, 1 - S); });
 
-        for(int p = 0; p<value_size; ++p){
+        for(size_t p = 0; p < value_size; ++p){
           update_s[p] = log(sum_s[p] / sum_s2[p]);
           update_b[p] = log(sum_b[p] / sum_b2[p]);
         }
 
         for(int individual: curr_split.leaf_index->individuals){
-          for(int p = 0; p<value_size; ++p){
+          for(size_t p = 0; p < value_size; ++p){
             if(samples_X[individual][curr_split.split_coordinate-1] < curr_split.split_point){
               weights[individual][p] *= exp(-0.5 * samples_Y[individual][p] * update_s[p]);
             }else{
